extract round from game servlet and put in round servlet
