+++ src/main/java/com/google/songgame/servlets/RoundServlet.java
@@ -10,6 +10,17 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.util.ArrayList;
+import com.google.api.services.youtube.model.Video;
+import com.google.appengine.api.datastore.DatastoreService;
+import com.google.appengine.api.datastore.Entity;
+import com.google.appengine.api.datastore.EmbeddedEntity;
+import com.google.appengine.api.datastore.PreparedQuery;
+import com.google.appengine.api.datastore.Query;
+import com.google.appengine.api.datastore.Query.SortDirection;
+import com.google.songgame.data.YoutubeParser;
+import com.google.appengine.api.datastore.DatastoreServiceFactory;
+import com.google.appengine.api.datastore.FetchOptions;
 
 @WebServlet("/round")
 public final class RoundServlet extends HttpServlet {
@@ -21,20 +32,37 @@ public final class RoundServlet extends HttpServlet {
   private static final String PUSHER_ROUND_CHANNEL_NAME = "start-round";
   private static final String PUSHER_GAME_CHANNEL_NAME = "start-game";
   private Pusher pusher;
+  private Gson gson;
+  DatastoreService datastore;
 
   @Override
   public void init() {
     pusher = new Pusher(APP_ID, CLIENT_KEY, CLIENT_SECRET);
     pusher.setCluster("us2");
     pusher.setEncrypted(true);
+
+    gson = new Gson();
+    datastore = DatastoreServiceFactory.getDatastoreService();
   }
 
   @Override
-  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+  public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
     pusher.trigger(
         PUSHER_APPLICATION_NAME,
         PUSHER_GAME_CHANNEL_NAME,
         Collections.singletonMap("message", "Start Game"));
+  }
+
+  @Override
+  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    Entity game = getCurrentGame();
+    EmbeddedEntity currentRound = getNewRound(game);
+    game.setProperty("currentRound", currentRound);
+    datastore.put(game);
+    EmbeddedEntity currentVideo = (EmbeddedEntity) currentRound.getProperty("video");
+    String currentVideoId = (String) currentVideo.getProperty("videoId");
+    String json = new Gson().toJson(currentVideoId);
+    response.getWriter().println(json);
   }
 
   @Override
@@ -45,3 +73,51 @@ public final class RoundServlet extends HttpServlet {
         Collections.singletonMap("message", "Start Round"));
   }
 }
+
+  private Entity getCurrentGame() {
+    Query gameQuery = new Query("Game").addSort("creationTime", SortDirection.DESCENDING);
+    PreparedQuery result = datastore.prepare(gameQuery);
+    Entity currentGame = result.asList(FetchOptions.Builder.withLimit(1)).get(0);
+    return currentGame;
+  }
+
+  private EmbeddedEntity getNewRound(Entity game) {
+    ArrayList<String> playlist = (ArrayList<String>) game.getProperty("playlist");
+    EmbeddedEntity video = getVideoEntity(playlist);
+    EmbeddedEntity userGuessStatuses = createUserGuessStatuses();
+
+    EmbeddedEntity currentRound = new EmbeddedEntity();
+    currentRound.setProperty("video", video);
+    currentRound.setProperty("userGuessStatuses", userGuessStatuses);
+
+    return currentRound;
+  }
+
+  private EmbeddedEntity getVideoEntity(ArrayList<String> playlist) {
+    YoutubeParser parser = new YoutubeParser();
+    Video video = parser.getRandomVideoFromPlaylist(playlist);
+    String videoId = video.getId();
+    String videoTitle = video.getSnippet().getTitle();
+
+    EmbeddedEntity videoEntity = new EmbeddedEntity();
+    videoEntity.setProperty("videoId", videoId);
+    videoEntity.setProperty("title", videoTitle);
+
+    return videoEntity;
+  }
+
+  private EmbeddedEntity createUserGuessStatuses() {
+    EmbeddedEntity userGuessStatuses = new EmbeddedEntity();
+    // TODO: @salilnadkarni, add more specific query to only get users with correct roomId
+    Query query = new Query("User");
+    PreparedQuery results = datastore.prepare(query);
+
+    for (Entity user : results.asIterable()) {
+      String userId = (String) user.getProperty("userId");
+      userGuessStatuses.setProperty(userId, false);
+    }
+
+    return userGuessStatuses;
+  }
+
+
